--- 06_pjt/01_skeleton/movies/forms.py
@@ -1,8 +1,0 @@
-from django import forms
-from .models import Movie
-
-
-class MovieForm(forms.ModelForm):
-    class Meta:
-        model = Movie
-        fields = ('title', 'description')
