--- 06_pjt/01_skeleton/movies/models.py
@@ -1,9 +1,0 @@
-from django.db import models
-
-
-class Movie(models.Model):
-    title = models.CharField(max_length=20)
-    description = models.TextField()
-
-    def __str__(self):
-        return self.title
