+++ movies/urls.py
@@ -10,4 +10,7 @@ urlpatterns = [
     path('<int:pk>/', views.detail, name='detail'),
     path('<int:pk>/delete/', views.delete, name='delete'),
     path('<int:pk>/update/', views.update, name='update'),
+    path('<int:pk>/comments/', views.comments_create, name='comments_create'),
+    path('<int:movie_pk>/comments/<int:comment_pk>/delete/', views.comments_delete, name='comments_delete'),
+    path('<int:pk>/likes/', views.likes, name='likes'),
 ]
